+++ docs/source/model_doc/bart.rst
@@ -49,13 +49,6 @@ BartTokenizer
     :members:
 
 
-MBartTokenizer
-~~~~~~~~~~~~~~~~~~~~~
-
-.. autoclass:: transformers.MBartTokenizer
-    :members: build_inputs_with_special_tokens, prepare_seq2seq_batch
-
-
 
 BartModel
 ~~~~~~~~~~~~~
